% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missingvalue.R
\name{conversion_omit}
\alias{conversion_omit}
\title{Converts missing values to omit.}
\description{
Converts missing values to omit.
}
\section{Super class}{
\code{\link[e4ts:conversion_base]{e4ts::conversion_base}} -> \code{conversion_omit}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-conversion_omit-new}{\code{conversion_omit$new()}}
\item \href{#method-conversion_omit-clone}{\code{conversion_omit$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="e4ts" data-topic="conversion_base" data-id="convert"><a href='../../e4ts/html/conversion_base.html#method-conversion_base-convert'><code>e4ts::conversion_base$convert()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-conversion_omit-new"></a>}}
\if{latex}{\out{\hypertarget{method-conversion_omit-new}{}}}
\subsection{Method \code{new()}}{
initialize object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{conversion_omit$new(missing_values = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{missing_values}}{missing_values name character vector}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-conversion_omit-clone"></a>}}
\if{latex}{\out{\hypertarget{method-conversion_omit-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{conversion_omit$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
