% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missingvalue.R
\name{conversion_zero}
\alias{conversion_zero}
\title{Converts missing values to 0.}
\description{
Converts missing values to 0.
}
\section{Super class}{
\code{\link[e4ts:conversion_base]{e4ts::conversion_base}} -> \code{conversion_zero}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{conversion_zero$new()}}
\item \href{#method-clone}{\code{conversion_zero$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="e4ts" data-topic="conversion_base" data-id="convert">}\href{../../e4ts/html/conversion_base.html#method-convert}{\code{e4ts::conversion_base$convert()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{conversion_zero$new(missing_values = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{conversion_zero$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
