% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{front}
\alias{front}
\title{frontend}
\usage{
front(host = "0.0.0.0", port = 3000)
}
\arguments{
\item{host}{The IP address that Shiny should listen}

\item{port}{A port number that Shiny will listen}
}
\description{
frontend by shiny app.
}
