% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measure.R
\name{measure_count}
\alias{measure_count}
\title{measure class for count}
\description{
measure class for count.
}
\section{Super class}{
\code{\link[e4ts:measure_base]{e4ts::measure_base}} -> \code{measure_count}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-measure_count-new}{\code{measure_count$new()}}
\item \href{#method-measure_count-calc_core}{\code{measure_count$calc_core()}}
\item \href{#method-measure_count-add_core}{\code{measure_count$add_core()}}
\item \href{#method-measure_count-clone}{\code{measure_count$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="e4ts" data-topic="measure_base" data-id="add"><a href='../../e4ts/html/measure_base.html#method-measure_base-add'><code>e4ts::measure_base$add()</code></a></span></li>
<li><span class="pkg-link" data-pkg="e4ts" data-topic="measure_base" data-id="calc"><a href='../../e4ts/html/measure_base.html#method-measure_base-calc'><code>e4ts::measure_base$calc()</code></a></span></li>
<li><span class="pkg-link" data-pkg="e4ts" data-topic="measure_base" data-id="info"><a href='../../e4ts/html/measure_base.html#method-measure_base-info'><code>e4ts::measure_base$info()</code></a></span></li>
<li><span class="pkg-link" data-pkg="e4ts" data-topic="measure_base" data-id="init"><a href='../../e4ts/html/measure_base.html#method-measure_base-init'><code>e4ts::measure_base$init()</code></a></span></li>
<li><span class="pkg-link" data-pkg="e4ts" data-topic="measure_base" data-id="set_weight"><a href='../../e4ts/html/measure_base.html#method-measure_base-set_weight'><code>e4ts::measure_base$set_weight()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-measure_count-new"></a>}}
\if{latex}{\out{\hypertarget{method-measure_count-new}{}}}
\subsection{Method \code{new()}}{
initialize object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{measure_count$new(name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{using column name}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-measure_count-calc_core"></a>}}
\if{latex}{\out{\hypertarget{method-measure_count-calc_core}{}}}
\subsection{Method \code{calc_core()}}{
calculation for sum.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{measure_count$calc_core(data)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{data.table}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-measure_count-add_core"></a>}}
\if{latex}{\out{\hypertarget{method-measure_count-add_core}{}}}
\subsection{Method \code{add_core()}}{
Add up the data.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{measure_count$add_core(target)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{target}}{A measure class}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-measure_count-clone"></a>}}
\if{latex}{\out{\hypertarget{method-measure_count-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{measure_count$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
