% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measure.R
\name{measure_sum}
\alias{measure_sum}
\title{Master class for aggregate items}
\description{
When creating aggregate items, be sure to inherit from and implement this class.
}
\section{Super class}{
\code{\link[e4ts:measure_base]{e4ts::measure_base}} -> \code{measure_sum}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{measure_sum$new()}}
\item \href{#method-calc_core}{\code{measure_sum$calc_core()}}
\item \href{#method-add_core}{\code{measure_sum$add_core()}}
\item \href{#method-clone}{\code{measure_sum$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="e4ts" data-topic="measure_base" data-id="add">}\href{../../e4ts/html/measure_base.html#method-add}{\code{e4ts::measure_base$add()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="e4ts" data-topic="measure_base" data-id="calc">}\href{../../e4ts/html/measure_base.html#method-calc}{\code{e4ts::measure_base$calc()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="e4ts" data-topic="measure_base" data-id="info">}\href{../../e4ts/html/measure_base.html#method-info}{\code{e4ts::measure_base$info()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="e4ts" data-topic="measure_base" data-id="init">}\href{../../e4ts/html/measure_base.html#method-init}{\code{e4ts::measure_base$init()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{measure_sum$new(name)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-calc_core"></a>}}
\if{latex}{\out{\hypertarget{method-calc_core}{}}}
\subsection{Method \code{calc_core()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{measure_sum$calc_core(data)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add_core"></a>}}
\if{latex}{\out{\hypertarget{method-add_core}{}}}
\subsection{Method \code{add_core()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{measure_sum$add_core(target)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{measure_sum$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
