% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stattable.R
\name{stattable}
\alias{stattable}
\title{create stat table}
\usage{
stattable(df, dimensions, measure)
}
\arguments{
\item{df}{data.table}

\item{dimensions}{column names string vector}

\item{measure}{A measure class}
}
\description{
create stat table
}
