% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e4ts-package.R
\docType{package}
\name{e4ts-package}
\alias{e4ts}
\alias{e4ts-package}
\title{e4ts: engine for statistical tables using tidy datas}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: tsutomu nagano \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
