% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measure.R
\name{measure_base}
\alias{measure_base}
\title{Master class for aggregate items}
\description{
Master class for aggregate items

Master class for aggregate items
}
\details{
When creating aggregate items, be sure to inherit from and implement this class.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{name}}{First or full name of the person.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-init}{\code{measure_base$init()}}
\item \href{#method-info}{\code{measure_base$info()}}
\item \href{#method-calc}{\code{measure_base$calc()}}
\item \href{#method-add}{\code{measure_base$add()}}
\item \href{#method-clone}{\code{measure_base$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-init"></a>}}
\if{latex}{\out{\hypertarget{method-init}{}}}
\subsection{Method \code{init()}}{
Create a new person object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{measure_base$init()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A new \code{Person} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-info"></a>}}
\if{latex}{\out{\hypertarget{method-info}{}}}
\subsection{Method \code{info()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{measure_base$info()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-calc"></a>}}
\if{latex}{\out{\hypertarget{method-calc}{}}}
\subsection{Method \code{calc()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{measure_base$calc(data)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add"></a>}}
\if{latex}{\out{\hypertarget{method-add}{}}}
\subsection{Method \code{add()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{measure_base$add(target)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{measure_base$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
