% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stattable.R
\name{stattable}
\alias{stattable}
\title{create stat table}
\usage{
stattable(
  df,
  dimensions,
  measure,
  conv = conversion_zero$new(),
  weight = "",
  calc_total = FALSE
)
}
\arguments{
\item{df}{data.table}

\item{dimensions}{column names string vector}

\item{measure}{A measure class}

\item{conv}{missing values conversion method conversion class}

\item{weight}{weight name}

\item{calc_total}{calculate total bool}
}
\description{
create stat table
}
